+++ cmd/pimo/main.go
@@ -171,23 +171,8 @@ There is NO WARRANTY, to the extent permitted by law.`, version, commit, buildDa
 				}
 
 				parser.RegisterMapCallback(elementName, func(m map[string]string) (map[string]string, error) {
-					dictionary := make(map[string]any, len(m))
-					for k, v := range m {
-						dictionary[k] = v
-					}
-					transformedData, err := ctx.ExecuteMap(dictionary)
-					if err != nil {
-						return nil, err
-					}
-					result := make(map[string]string, len(m))
-					for k, v := range transformedData {
-						stringValue, ok := v.(string)
-						if !ok {
-							return nil, fmt.Errorf("Result is not a string")
-						}
-						result[k] = stringValue
-					}
-					return result, nil
+					newList, _ := pimo.XMLCallback(ctx, m)
+					return newList, nil
 				})
 			}
 			err := parser.Stream()